requirements.txt
# python3 -m pip install --user -U -r requirements.txt
rootpath
findspark
pyspark
apache-airflow==1.10.12